[Upd] Moved calendar view to teamname app
