+++ ctf/views.py
@@ -6,7 +6,7 @@ from django.shortcuts import get_object_or_404
 from django.contrib.auth.decorators import login_required
 
 from ctf.models import CTF, Challenge, ChallengeFile
-from ctf.forms import ChallengeForm, CTForm
+from ctf.forms import ChallengeForm
 
 
 @login_required
@@ -46,16 +46,6 @@ def edit_chall(request, ctf_id, chall_id):
         "chall_id": chall_id,
     }
     return render(request, "form.html", context)
-
-
-@login_required
-def calendar(request):
-    context = {
-        "page_title": "Calendar",
-        "events": CTF.objects.all(),
-        "form": CTForm(),
-    }
-    return render(request, "calendar.html", context)
 
 
 @login_required
